Add property of samples per pixel
